+++ src/test/kotlin/com/zepben/evolve/streaming/grpc/ConnectTest.kt
@@ -18,7 +18,6 @@ import org.hamcrest.Matchers.equalTo
 import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
-import java.net.http.HttpClient
 
 internal class ConnectTest {
 
@@ -117,11 +116,7 @@ internal class ConnectTest {
     fun connectWithSecretWithKnownTokenFetcherConfig() {
         mockkConstructor(ZepbenTokenFetcher::class)
         every {
-            constructedWith<ZepbenTokenFetcher>(
-                EqMatcher("audience"), EqMatcher("issuerDomain"),  EqMatcher(AuthMethod.OAUTH),
-                AllAnyMatcher<String>(), AllAnyMatcher<String>(), AllAnyMatcher<JsonObject>(), AllAnyMatcher<JsonObject>(),
-                AllAnyMatcher<String?>(), AllAnyMatcher<String>(), AllAnyMatcher<(JsonObject) -> String>()
-            ).tokenRequestData
+            anyConstructed<ZepbenTokenFetcher>().tokenRequestData
         } returns tokenRequestData
 
         every { gcbWithTls.withTokenFetcher(any()) } returns gcbWithAuth
@@ -208,11 +203,7 @@ internal class ConnectTest {
     fun connectWithPasswordWithKnownTokenFetcherConfig() {
         mockkConstructor(ZepbenTokenFetcher::class)
         every {
-            constructedWith<ZepbenTokenFetcher>(
-                EqMatcher("audience"), EqMatcher("issuerDomain"),  EqMatcher(AuthMethod.OAUTH),
-                AllAnyMatcher<String>(), AllAnyMatcher<String>(), AllAnyMatcher<JsonObject>(), AllAnyMatcher<JsonObject>(),
-                AllAnyMatcher<String?>(), AllAnyMatcher<String>(), AllAnyMatcher<(JsonObject) -> String>()
-            ).tokenRequestData
+            anyConstructed<ZepbenTokenFetcher>().tokenRequestData
         } returns tokenRequestData
 
         every { gcbWithTls.withTokenFetcher(any()) } returns gcbWithAuth